+++ test/specs/options/headers.spec.js
@@ -6,28 +6,29 @@ describe('headers (option)', () => {
     expect(picker.options.headers).to.be.false;
   });
 
-  it('should show column headers', () => {
+  it('should show the column headers', () => {
     const input = window.createInput();
     const picker = new Picker(input, {
       headers: true,
     });
 
-    expect(picker.picker.querySelector('.picker-years').dataset.header).to.equal('Year');
-    expect(picker.picker.querySelector('.picker-months').dataset.header).to.equal('Month');
-    expect(picker.picker.querySelector('.picker-days').dataset.header).to.equal('Day');
+    expect(picker.picker.querySelector('.picker-years .picker-cell__header').textContent).to.equal('Year');
+    expect(picker.picker.querySelector('.picker-months .picker-cell__header').textContent).to.equal('Month');
+    expect(picker.picker.querySelector('.picker-days .picker-cell__header').textContent).to.equal('Day');
   });
 
   it('should show the custom column headers', () => {
     const input = window.createInput();
     const picker = new Picker(input, {
-      headers: {
+      headers: true,
+      text: {
         year: '年',
         month: '月',
       },
     });
 
-    expect(picker.picker.querySelector('.picker-years').dataset.header).to.equal('年');
-    expect(picker.picker.querySelector('.picker-months').dataset.header).to.equal('月');
-    expect(picker.picker.querySelector('.picker-days').dataset.header).to.equal('Day');
+    expect(picker.picker.querySelector('.picker-years .picker-cell__header').textContent).to.equal('年');
+    expect(picker.picker.querySelector('.picker-months .picker-cell__header').textContent).to.equal('月');
+    expect(picker.picker.querySelector('.picker-days .picker-cell__header').textContent).to.equal('Day');
   });
 });